requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
diff-match-patch==20230430
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
django-import-export==4.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
importlib-metadata==6.8.0
kombu==5.3.2
Markdown==3.4.4
numpy==1.25.2
pandas==2.1.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.8
zipp==3.16.2
pusher
dotenv
drf-spectacular
resend